+++ test/Certes.Tests/AcmeClientTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Net;
 using System.Net.Http;
+using System.Reflection;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
@@ -186,6 +187,147 @@ namespace Certes
             }
         }
 
+        [Fact]
+        public void CanCreateWithUri()
+        {
+            var uri = WellKnownServers.LetsEncryptStaging;
+
+            using (var client = new AcmeClient(uri))
+            {
+                var type = typeof(AcmeClient).GetTypeInfo();
+                var field = type.GetField("shouldDisposeHander", BindingFlags.NonPublic | BindingFlags.Instance);
+                Assert.True((bool)field.GetValue(client));
+                Assert.NotNull(client.HttpHandler);
+
+                Assert.Equal(uri, client.HttpHandler.ServerUri);
+            }
+        }
+
+        [Fact]
+        public async Task CanHandlerxistingAuthorization()
+        {
+            var accountKey = await Helper.LoadkeyV1();
+            var authzUri = new Uri("http://example.com/new-authz");
+            var authzLoc = new Uri("http://example.com/authz/111");
+            var mock = MockHttp(async req =>
+            {
+                if (req.Method == HttpMethod.Post && req.RequestUri == authzUri)
+                {
+                    var payload = await ParsePayload<AuthorizationEntity>(req);
+                    return CreateResponse(null, HttpStatusCode.SeeOther, authzLoc);
+                }
+
+                if (req.Method == HttpMethod.Get && req.RequestUri == authzLoc)
+                {
+                    return CreateResponse(new AuthorizationEntity
+                    {
+                        Identifier = new AuthorizationIdentifier
+                        {
+                            Type = AuthorizationIdentifierTypes.Dns,
+                            Value = "www.example.com",
+                        },
+                        Status = EntityStatus.Pending,
+                    }, HttpStatusCode.OK, authzLoc);
+                }
+
+                return null;
+            });
+
+            using (var http = new HttpClient(mock.Object))
+            using (var handler = new AcmeHttpHandler(server, http))
+            {
+                using (var client = new AcmeClient(handler))
+                {
+                    client.Use(accountKey.Export());
+
+                    var authz = await client.NewAuthorization(new AuthorizationIdentifier
+                    {
+                        Type = AuthorizationIdentifierTypes.Dns,
+                        Value = "www.example.com"
+                    });
+
+                    Assert.NotNull(authz);
+                    Assert.Equal(authzLoc, authz.Location);
+                }
+
+                mock.Protected().Verify("Dispose", Times.Never(), true);
+            }
+        }
+
+        [Fact]
+        public async Task CanNewRegistraton()
+        {
+            var accountKey = await Helper.LoadkeyV1();
+            var contacts = new string[] { "mailto:user@example.com" };
+            var regLocation = new Uri("http://example.com/reg/1");
+            var mock = MockHttp(async req =>
+            {
+                if (req.Method == HttpMethod.Post && req.RequestUri == Helper.MockDirectoryV1.NewReg)
+                {
+                    var payload = await ParsePayload<RegistrationEntity>(req);
+                    Assert.Equal(ResourceTypes.NewRegistration, payload.Resource);
+                    Assert.Equal(contacts.Clone(), payload.Contact);
+
+                    var respJson = new
+                    {
+                        contact = payload.Contact,
+                        agreement = payload.Agreement,
+                        resource = ResourceTypes.Registration
+                    };
+
+                    var resp = CreateResponse(respJson, HttpStatusCode.Created, regLocation);
+                    resp.Headers.Add("Link", $"<{tos}>; rel=\"terms-of-service\"");
+                    return resp;
+                }
+
+                return null;
+            });
+
+            using (var http = new HttpClient(mock.Object))
+            using (var handler = new AcmeHttpHandler(server, http))
+            {
+                using (var client = new AcmeClient(handler))
+                {
+                    client.Use(accountKey.Export());
+                    var result = await client.NewRegistraton(contacts);
+                    Assert.Equal(ResourceTypes.Registration, result.Data.Resource);
+                    Assert.Equal(regLocation, result.Location);
+                }
+
+                mock.Protected().Verify("Dispose", Times.Never(), true);
+            }
+        }
+
+        [Fact]
+        public async Task ShouldFailWhenNewAuthorizationWithoutAccount()
+        {
+            using (var client = new AcmeClient(WellKnownServers.LetsEncryptStaging))
+            {
+                await Assert.ThrowsAsync<InvalidOperationException>(
+                    () => client.NewAuthorization(new AuthorizationIdentifier()));
+            }
+        }
+
+        [Fact]
+        public async Task ShouldFailWhenUpdateRegistrationWithoutAccount()
+        {
+            using (var client = new AcmeClient(WellKnownServers.LetsEncryptStaging))
+            {
+                await Assert.ThrowsAsync<InvalidOperationException>(
+                    () => client.UpdateRegistration(new AcmeAccount()));
+            }
+        }
+
+        [Fact]
+        public async Task ShouldFailWhenCompleteChallengeWithoutAccount()
+        {
+            using (var client = new AcmeClient(WellKnownServers.LetsEncryptStaging))
+            {
+                await Assert.ThrowsAsync<InvalidOperationException>(
+                    () => client.CompleteChallenge(new ChallengeEntity()));
+            }
+        }
+
         private async Task<T> ParsePayload<T>(HttpRequestMessage message)
         {
             var accountKey = await Helper.LoadkeyV1();
@@ -215,7 +357,7 @@ namespace Certes
 
         private Mock<HttpMessageHandler> MockHttp(Func<HttpRequestMessage, Task<HttpResponseMessage>> provider)
         {
-            var mock = new Mock<HttpMessageHandler>();
+            var mock = new Mock<HttpMessageHandler>(MockBehavior.Strict);
 
             mock.Protected()
                 .Setup<Task<HttpResponseMessage>>("SendAsync", ItExpr.IsAny<HttpRequestMessage>(), ItExpr.IsAny<CancellationToken>())
@@ -239,6 +381,8 @@ namespace Certes
                     }
                 });
 
+            mock.Protected().Setup("Dispose", true);
+
             return mock;
         }
 
@@ -246,6 +390,7 @@ namespace Certes
         {
             var resp = new HttpResponseMessage();
             resp.Headers.Location = location;
+            resp.StatusCode = statusCode;
             if (payload != null)
             {
                 resp.Content = new StringContent(JsonConvert.SerializeObject(payload), Encoding.UTF8, JsonContentType);