+++ test/Certes.Tests/AcmeRequestExceptionTests.cs
@@ -50,11 +50,11 @@ namespace Certes
                 Assert.Equal("certes", deserialized.Message);
             }
         }
-
+        
         [Fact]
-        public void CanSerializeWithError()
+        public void CanSerializeWithErrorAndMessage()
         {
-            var error = new AcmeError { Detail = "error" };
+            var error = new AcmeError { Type = "t", Detail = "error" };
             var ex = new AcmeRequestException("certes", error);
 
             var serializer = new BinaryFormatter();
@@ -66,6 +66,7 @@ namespace Certes
                 buffer.Seek(0, SeekOrigin.Begin);
                 var deserialized = (AcmeRequestException)serializer.Deserialize(buffer);
 
+                Assert.Equal("certes\nt: error", deserialized.Message);
                 Assert.NotNull(deserialized.Error.Detail);
                 Assert.Equal("error", deserialized.Error.Detail);
             }