+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Support `tls-alpn-01` challenge ([#125][i125])
+- aDD `ACME` error details to exception messages ([#109][i109])
 
 ## [2.2.2] - 2018-05-31
 ### Changed
@@ -96,3 +97,4 @@ All notable changes to this project will be documented in this file.
 [i106]: https://github.com/fszlin/certes/issues/106
 [i112]: https://github.com/fszlin/certes/issues/112
 [i125]: https://github.com/fszlin/certes/issues/125
+[i109]: https://github.com/fszlin/certes/issues/109